refactor db interaction
