+++ analytics/app/dashboard.py
@@ -2,7 +2,6 @@ import dash
 from dash import html, dcc, Input, Output
 from flask import send_from_directory
 import pandas as pd
-import sqlite3
 import os
 import sys
 import logging
@@ -11,7 +10,8 @@ import logging
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../..")))
 import src.config as config
 
-# Import diagram modules
+# Import data and diagram modules
+from .data import get_session_data, get_top_images
 from .diagrams_usage import (
     create_sessions_timeline, create_os_chart, create_browser_chart,
     create_mobile_pie, create_generation_status_chart
@@ -21,7 +21,7 @@ from .diagrams_origin import (
     create_country_chart, create_city_chart
 )
 from .diagrams_uploads import (
-    get_top_images, create_image_uploads_timeline,
+    create_image_uploads_timeline,
     create_top_images_chart
 )
 
@@ -42,66 +42,6 @@ console_handler.setFormatter(
 )
 logger.addHandler(console_handler)
 
-def get_session_data(start_date=None, end_date=None, include_generation_status=False, include_input_data=False):
-    """Get session data from analytics database.
-    
-    Args:
-        start_date (str, optional): Start date in YYYY-MM-DD format
-        end_date (str, optional): End date in YYYY-MM-DD format
-    """
-    config.read_configuration()
-    connection = sqlite3.connect(config.get_analytics_db_path())
-    
-    # Handle datetime format YYYY-MM-DD HH:MM:SS
-    base_query = """
-    SELECT 
-        s.*,
-        date(s.Timestamp) as Date
-    """
-    
-    if include_generation_status:
-        base_query += """,
-        CASE 
-            WHEN EXISTS (
-                SELECT 1 FROM tblGenerations g WHERE g.Session = s.Session
-                UNION
-                SELECT 1 FROM tblInput i WHERE i.Session = s.Session
-            ) THEN 1
-            ELSE 0
-        END as HasStartedGeneration
-        """
-    
-    if include_input_data:
-        base_query += """,
-        (
-            SELECT COUNT(*)
-            FROM tblInput i
-            WHERE i.Session = s.Session
-        ) as ImageUploads
-        """
-    
-    base_query += " FROM tblSessions s"
-    
-    if start_date and end_date:
-        base_query += " WHERE date(s.Timestamp) BETWEEN ? AND ?"
-    
-    base_query += " ORDER BY s.Timestamp"
-    
-    query = base_query
-    
-    params = [start_date, end_date] if start_date and end_date else []
-    
-    df = pd.read_sql_query(query, connection, params=params)
-    connection.close()
-    
-    # Log data loading info
-    logger.info(f"Data loaded: {len(df)} rows")
-    if len(df) > 0:
-        logger.debug(f"Date range: {df['Date'].min()} to {df['Date'].max()}")
-        logger.debug(f"Sample dates: {df['Date'].head().tolist()}")
-    
-    return df
-
 # Initialize the Dash app
 app = dash.Dash(__name__, title="AI Anime Maker Analytics")
 
@@ -120,8 +60,13 @@ initial_end_date = pd.Timestamp.now().strftime('%Y-%m-%d')
 initial_start_date = (pd.Timestamp.now() - pd.Timedelta(days=30)).strftime('%Y-%m-%d')
 
 # Get initial data
-df = get_session_data(initial_start_date, initial_end_date, include_generation_status=True, include_input_data=True)
-top_images_df = get_top_images()
+df = get_session_data(
+    initial_start_date, 
+    initial_end_date, 
+    include_generation_status=True, 
+    include_input_data=True
+)
+top_images_df = get_top_images(initial_start_date, initial_end_date)
 
 # Create layout
 app.layout = html.Div(style=LAYOUT_STYLE, children=[
@@ -289,8 +234,13 @@ app.layout = html.Div(style=LAYOUT_STYLE, children=[
 )
 def update_charts(start_date, end_date):
     """Update all charts based on selected date range."""
-    df = get_session_data(start_date, end_date, include_generation_status=True, include_input_data=True)
-    top_images_df = get_top_images()
+    df = get_session_data(
+        start_date, 
+        end_date, 
+        include_generation_status=True, 
+        include_input_data=True
+    )
+    top_images_df = get_top_images(start_date, end_date)
     return [
         create_sessions_timeline(df),
         create_os_chart(df),
